增加了树形弹出框参照

+++ src/rcplatform/src/main/java/com/ruanchuangsoft/platform/service/impl/TransboxmainServiceImpl.java
@@ -1,7 +1,5 @@
 package com.ruanchuangsoft.platform.service.impl;
 
-import com.ruanchuangsoft.platform.dao.TransboxdetailDao;
-import com.ruanchuangsoft.platform.entity.TransboxdetailEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,7 +7,9 @@ import java.util.List;
 import java.util.Map;
 
 import com.ruanchuangsoft.platform.dao.TransboxmainDao;
+import com.ruanchuangsoft.platform.dao.TransboxdetailDao;
 import com.ruanchuangsoft.platform.entity.TransboxmainEntity;
+import com.ruanchuangsoft.platform.entity.TransboxdetailEntity;
 import com.ruanchuangsoft.platform.service.TransboxmainService;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,10 +18,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class TransboxmainServiceImpl implements TransboxmainService {
 	@Autowired
 	private TransboxmainDao transboxmainDao;
-	@Autowired
-	private TransboxdetailDao transboxdetailDao;
-	
-	@Override
+
+    @Autowired
+    private TransboxdetailDao transboxdetailDao;
+
+
+    @Override
 	public TransboxmainEntity queryObject(Long id){
 		return transboxmainDao.queryObject(id);
 	}
@@ -38,26 +40,44 @@ public class TransboxmainServiceImpl implements TransboxmainService {
 	
 	@Override
 	public void save(TransboxmainEntity transboxmain){
+        for(TransboxdetailEntity item:transboxmain.getDetails()){
+				transboxdetailDao.save(item);
+        }
 		transboxmainDao.save(transboxmain);
-		for(TransboxdetailEntity item:transboxmain.getDetails()){
-			transboxdetailDao.save(item);
-		}
 	}
 	
 	@Override
 	public void update(TransboxmainEntity transboxmain){
-		transboxmainDao.update(transboxmain);
-
+        for(TransboxdetailEntity item:transboxmain.getDetails()){
+				transboxdetailDao.save(item);
+        }
+			transboxmainDao.update(transboxmain);
 	}
 	
 	@Override
 	public void delete(Long id){
-		transboxmainDao.delete(id);
+		 TransboxmainEntity main = queryObject(id);
+        if (main != null) {
+				transboxmainDao.delete(id);
+
+				transboxdetailDao.deleteByBillNo(main.getBillno());
+
+        }
+
 	}
 	
 	@Override
-	public void deleteBatch(Long[] ids){
-		transboxmainDao.deleteBatch(ids);
-	}
-	
+    public void deleteBatch(Long[] ids){
+			transboxmainDao.deleteBatch(ids);
+    }
+
+    @Override
+    public void auditBatch(Long[] ids){
+			transboxmainDao.auditBatch(ids);
+    }
+
+    @Override
+    public void unauditBatch(Long[] ids){
+			transboxmainDao.unauditBatch(ids);
+    }
 }
